reattach time: truncated exponential instead of uniform + some bugs fixed
